+++ src/vs/platform/userDataSync/test/common/userDataSyncClient.ts
@@ -43,8 +43,8 @@ import { UserDataSyncService } from 'vs/platform/userDataSync/common/userDataSyn
 import { UserDataSyncStoreManagementService, UserDataSyncStoreService } from 'vs/platform/userDataSync/common/userDataSyncStoreService';
 import { IUserDataProfile, IUserDataProfilesService, UserDataProfilesService } from 'vs/platform/userDataProfile/common/userDataProfile';
 import { NullPolicyService } from 'vs/platform/policy/common/policy';
-import { IProfileStorageService } from 'vs/platform/storage/common/profileStorageService';
-import { TestProfileStorageService } from 'vs/platform/storage/test/common/profileStorageService.test';
+import { IUserDataSyncProfilesStorageService } from 'vs/platform/userDataSync/common/userDataSyncProfilesStorageService';
+import { TestUserDataSyncProfilesStorageService } from 'vs/platform/userDataSync/test/common/userDataSyncProfilesStorageService.test';
 
 export class UserDataSyncClient extends Disposable {
 
@@ -92,7 +92,7 @@ export class UserDataSyncClient extends Disposable {
 
 		const storageService = new TestStorageService(userDataProfilesService.defaultProfile);
 		this.instantiationService.stub(IStorageService, this._register(storageService));
-		this.instantiationService.stub(IProfileStorageService, this._register(new TestProfileStorageService(storageService)));
+		this.instantiationService.stub(IUserDataSyncProfilesStorageService, this._register(new TestUserDataSyncProfilesStorageService(storageService)));
 
 		const configurationService = this._register(new ConfigurationService(userDataProfilesService.defaultProfile.settingsResource, fileService, new NullPolicyService(), logService));
 		await configurationService.initialize();
@@ -307,7 +307,7 @@ class TestStorageService extends InMemoryStorageService {
 	constructor(private readonly profileStorageProfile: IUserDataProfile) {
 		super();
 	}
-	override getProfileStorageProfile(): IUserDataProfile | undefined {
-		return this.profileStorageProfile;
+	override isProfileStorageFor(profile: IUserDataProfile): boolean {
+		return this.profileStorageProfile.id === profile.id;
 	}
 }
